backend/requirements.txt
<<<<<<< HEAD
# Core
Flask
Flask-CORS

# Binance API
ccxt

# Data Analysis
pandas
numpy

# Environment
python-dotenv

# HTTP Requests
requests

# Date/Time
=======
# Core
Flask
Flask-CORS

# Binance API
ccxt

# Data Analysis
pandas
numpy

# Environment
python-dotenv

# HTTP Requests
requests

# Date/Time
>>>>>>> b599aa4371431de9ca4e937d9a077375f81f08b2
python-dateutil